Moves search and improves code
