+++ mochi_code/commands/exceptions.py
@@ -1,0 +1,5 @@
+"""Exceptions for the commands module."""
+
+
+class MochiCannotContinue(Exception):
+    """Raised when mochi cannot continue with the current command."""
